+++ services/gateway/src/routes/create_account.rs
@@ -6,7 +6,7 @@ use crate::{
 };
 use axum::{Extension, Json, extract::State};
 use tracing::instrument;
-use world_id_core::api_types::{CreateAccountRequest, GatewayStatusResponse};
+use world_id_primitives::api_types::{CreateAccountRequest, GatewayStatusResponse};
 
 /// POST /create-account
 ///