+++ services/indexer/tests/test_blockchain_streams.rs
@@ -9,9 +9,10 @@ use alloy::{
     rpc::types::Filter,
     signers::local::PrivateKeySigner,
 };
+use eddsa_babyjubjub::EdDSAPrivateKey;
 use futures_util::{StreamExt, TryStreamExt};
-use world_id_core::{EdDSAPrivateKey, world_id_registry::WorldIdRegistry};
 use world_id_indexer::blockchain::{Blockchain, BlockchainError, BlockchainEvent, RegistryEvent};
+use world_id_registries::world_id::WorldIdRegistry;
 use world_id_services_common::ProviderArgs;
 use world_id_test_utils::anvil::TestAnvil;