+++ services/oprf-node/src/lib.rs
@@ -31,7 +31,6 @@ use std::sync::Arc;
 use ark_bn254::Bn254;
 use circom_types::groth16::VerificationKey;
 use eyre::Context;
-use taceo_nodes_common::web3;
 use taceo_oprf::service::{
     OprfServiceBuilder, StartedServices, secret_manager::SecretManagerService,
 };
@@ -59,13 +58,6 @@ pub mod metrics;
 #[allow(clippy::struct_field_names, reason = "Has the watcher suffix in name")]
 pub struct WorldOprfNodeTasks {
     key_event_watcher: tokio::task::JoinHandle<eyre::Result<()>>,
-    merkle_watcher: tokio::task::JoinHandle<eyre::Result<()>>,
-    rp_registry_watcher: tokio::task::JoinHandle<eyre::Result<()>>,
-    schema_issuer_registry_watcher: tokio::task::JoinHandle<eyre::Result<()>>,
-    // We also store the RpcProvider here.
-    //
-    // We want it to live at least as long as the sub-tasks need to complete their graceful shutdown.
-    _rpc_provider: web3::RpcProvider,
 }
 
 impl WorldOprfNodeTasks {
@@ -80,21 +72,7 @@ impl WorldOprfNodeTasks {
     /// - any task returns an error, or
     /// - any task panics or is aborted.
     pub async fn join(self) -> eyre::Result<()> {
-        let (
-            key_event_watcher,
-            merkle_watcher,
-            rp_registry_watcher,
-            schema_issuer_registry_watcher,
-        ) = tokio::join!(
-            self.key_event_watcher,
-            self.merkle_watcher,
-            self.rp_registry_watcher,
-            self.schema_issuer_registry_watcher
-        );
-        key_event_watcher??;
-        merkle_watcher??;
-        rp_registry_watcher??;
-        schema_issuer_registry_watcher??;
+        self.key_event_watcher.await??;
         Ok(())
     }
 }
@@ -117,7 +95,7 @@ impl WorldOprfNodeTasks {
 /// - Constructing the Axum router for handling incoming HTTP requests
 ///
 /// The returned [`WorldOprfNodeTasks`] contains all long-running background
-/// tasks (watchers and key event handling).
+/// tasks (key event handling).
 ///
 /// # Arguments
 /// - `config`: Full node configuration.
@@ -148,37 +126,26 @@ pub async fn start(
     let started_services = StartedServices::default();
 
     tracing::info!("connecting to RPC..");
-    let rpc_provider =
-        taceo_nodes_common::web3::RpcProviderBuilder::with_config(&config.rpc_provider_config)
+    let http_rpc_provider =
+        taceo_nodes_common::web3::HttpRpcProviderBuilder::with_config(&config.rpc_provider_config)
             .environment(node_config.environment)
             .build()
-            .await
             .context("while init blockchain connection")?;
 
     tracing::info!("init merkle watcher..");
-    let (merkle_watcher, merkle_watcher_task) = MerkleWatcher::init(
+    let merkle_watcher = MerkleWatcher::init(
         config.world_id_registry_contract,
-        &rpc_provider,
-        config.max_merkle_cache_size,
-        config.cache_maintenance_interval,
-        started_services.new_service(),
-        cancellation_token.clone(),
-    )
-    .await
-    .context("while starting merkle watcher")?;
+        &http_rpc_provider,
+        config.merkle_cache_config,
+    );
 
     tracing::info!("init RpRegistry watcher..");
-    let (rp_registry_watcher, rp_registry_watcher_task) = RpRegistryWatcher::init(
+    let rp_registry_watcher = RpRegistryWatcher::init(
         config.rp_registry_contract,
-        rpc_provider.clone(),
-        config.rp_cache_config,
-        config.cache_maintenance_interval,
+        http_rpc_provider.clone(),
         config.timeout_external_eth_call,
-        started_services.new_service(),
-        cancellation_token.clone(),
-    )
-    .await
-    .context("while starting rp registry watcher")?;
+        config.rp_cache_config,
+    );
 
     let query_vk = serde_json::from_str::<VerificationKey<Bn254>>(QUERY_VERIFICATION_KEY)
         .expect("can deserialize embedded vk");
@@ -191,11 +158,10 @@ pub async fn start(
         NonceHistory::init(
             // keep cache for 2x so that we catch all replays that would be valid and some that would be invalid anyways
             config.current_time_stamp_max_difference * 2,
-            config.cache_maintenance_interval,
         ),
         config.current_time_stamp_max_difference,
         config.timeout_external_eth_call,
-        rpc_provider.clone(),
+        http_rpc_provider.clone(),
         Arc::clone(&query_vk),
     ));
 
@@ -208,26 +174,19 @@ pub async fn start(
         NonceHistory::init(
             // keep cache for 2x so that we catch all replays that would be valid and some that would be invalid anyways
             config.current_time_stamp_max_difference * 2,
-            config.cache_maintenance_interval,
         ),
         config.current_time_stamp_max_difference,
         config.timeout_external_eth_call,
-        rpc_provider.clone(),
+        http_rpc_provider.clone(),
         Arc::clone(&query_vk),
     ));
 
     tracing::info!("init CredentialSchemaIssuerRegistry watcher..");
-    let (schema_issuer_registry_watcher, schema_issuer_registry_watcher_task) =
-        SchemaIssuerRegistryWatcher::init(
-            config.credential_schema_issuer_registry_contract,
-            &rpc_provider,
-            config.issuer_cache_config,
-            config.cache_maintenance_interval,
-            started_services.new_service(),
-            cancellation_token.clone(),
-        )
-        .await
-        .context("while starting schema issuer registry watcher")?;
+    let schema_issuer_registry_watcher = SchemaIssuerRegistryWatcher::init(
+        config.credential_schema_issuer_registry_contract,
+        &http_rpc_provider,
+        config.issuer_cache_config,
+    );
 
     tracing::info!("init credential blinding factor oprf request auth service..");
     let credential_blinding_factor_oprf_req_auth_service =
@@ -241,7 +200,7 @@ pub async fn start(
     let (router, key_event_watcher) = OprfServiceBuilder::init(
         node_config,
         secret_manager,
-        rpc_provider.clone(),
+        http_rpc_provider.clone(),
         started_services,
         cancellation_token.clone(),
     )
@@ -259,13 +218,7 @@ pub async fn start(
         session_oprf_req_auth_service,
     )
     .build();
-    let tasks = WorldOprfNodeTasks {
-        key_event_watcher,
-        merkle_watcher: merkle_watcher_task,
-        rp_registry_watcher: rp_registry_watcher_task,
-        schema_issuer_registry_watcher: schema_issuer_registry_watcher_task,
-        _rpc_provider: rpc_provider,
-    };
+    let tasks = WorldOprfNodeTasks { key_event_watcher };
 
     Ok((router, tasks))
 }